Revert "feat: Multi-calendar support in resource tables and analytics engine [2.39-DHIS2-15531]"
