+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/period/PeriodType.java
@@ -130,7 +130,6 @@ public abstract class PeriodType implements Serializable {
           new MonthlyPeriodType(),
           new BiMonthlyPeriodType(),
           new QuarterlyPeriodType(),
-          new QuarterlyNovemberPeriodType(),
           new SixMonthlyPeriodType(),
           new SixMonthlyAprilPeriodType(),
           new SixMonthlyNovemberPeriodType(),
