+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/period/PeriodTypeEnum.java
@@ -41,7 +41,6 @@ public enum PeriodTypeEnum {
   FINANCIAL_OCT("FinancialOct"),
   MONTHLY("Monthly"),
   QUARTERLY("Quarterly"),
-  QUARTERLY_NOV("QuarterlyNov"),
   SIX_MONTHLY_APRIL("SixMonthlyApril"),
   SIX_MONTHLY_NOV("SixMonthlyNov"),
   SIX_MONTHLY("SixMonthly"),
