+++ dhis-2/dhis-services/dhis-service-administration/src/main/java/org/hisp/dhis/resourcetable/DefaultResourceTableService.java
@@ -52,7 +52,6 @@ import org.hisp.dhis.jdbc.StatementBuilder;
 import org.hisp.dhis.organisationunit.OrganisationUnitGroupSet;
 import org.hisp.dhis.organisationunit.OrganisationUnitLevel;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.period.PeriodDataProvider;
 import org.hisp.dhis.period.PeriodService;
 import org.hisp.dhis.resourcetable.table.CategoryOptionComboNameResourceTable;
 import org.hisp.dhis.resourcetable.table.CategoryOptionComboResourceTable;
@@ -97,8 +96,6 @@ public class DefaultResourceTableService implements ResourceTableService {
   private final StatementBuilder statementBuilder;
 
   private final AnalyticsExportSettings analyticsExportSettings;
-
-  private final PeriodDataProvider periodDataProvider;
 
   @Override
   @Transactional
@@ -181,8 +178,7 @@ public class DefaultResourceTableService implements ResourceTableService {
   @Override
   public void generateDatePeriodTable() {
     resourceTableStore.generateResourceTable(
-        new DatePeriodResourceTable(
-            periodDataProvider.getAvailableYears(), analyticsExportSettings.getTableType()));
+        new DatePeriodResourceTable(null, analyticsExportSettings.getTableType()));
   }
 
   @Override
