+++ dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/table/AbstractJdbcTableManager.java
@@ -57,6 +57,7 @@ import org.hisp.dhis.analytics.AnalyticsTablePhase;
 import org.hisp.dhis.analytics.AnalyticsTableType;
 import org.hisp.dhis.analytics.AnalyticsTableUpdateParams;
 import org.hisp.dhis.analytics.partition.PartitionManager;
+import org.hisp.dhis.calendar.Calendar;
 import org.hisp.dhis.category.CategoryService;
 import org.hisp.dhis.common.IdentifiableObjectManager;
 import org.hisp.dhis.common.ValueType;
@@ -69,7 +70,6 @@ import org.hisp.dhis.jdbc.StatementBuilder;
 import org.hisp.dhis.organisationunit.OrganisationUnitGroupSet;
 import org.hisp.dhis.organisationunit.OrganisationUnitLevel;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.period.PeriodDataProvider;
 import org.hisp.dhis.period.PeriodType;
 import org.hisp.dhis.resourcetable.ResourceTableService;
 import org.hisp.dhis.setting.SettingKey;
@@ -131,8 +131,6 @@ public abstract class AbstractJdbcTableManager implements AnalyticsTableManager 
   protected final JdbcTemplate jdbcTemplate;
 
   protected final AnalyticsExportSettings analyticsExportSettings;
-
-  protected final PeriodDataProvider periodDataProvider;
 
   private static final String WITH_AUTOVACUUM_ENABLED_FALSE = "with(autovacuum_enabled = false)";
 
@@ -410,6 +408,7 @@ public abstract class AbstractJdbcTableManager implements AnalyticsTableManager 
       List<Integer> dataYears,
       List<AnalyticsTableColumn> dimensionColumns,
       List<AnalyticsTableColumn> valueColumns) {
+    Calendar calendar = PeriodType.getCalendar();
 
     Collections.sort(dataYears);
 
@@ -418,7 +417,9 @@ public abstract class AbstractJdbcTableManager implements AnalyticsTableManager 
 
     for (Integer year : dataYears) {
       table.addPartitionTable(
-          year, PartitionUtils.getStartDate(year), PartitionUtils.getEndDate(year));
+          year,
+          PartitionUtils.getStartDate(calendar, year),
+          PartitionUtils.getEndDate(calendar, year));
     }
 
     return table;
