+++ dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/table/JdbcCompletenessTargetTableManager.java
@@ -60,7 +60,6 @@ import org.hisp.dhis.jdbc.StatementBuilder;
 import org.hisp.dhis.organisationunit.OrganisationUnitGroupSet;
 import org.hisp.dhis.organisationunit.OrganisationUnitLevel;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.period.PeriodDataProvider;
 import org.hisp.dhis.resourcetable.ResourceTableService;
 import org.hisp.dhis.setting.SystemSettingManager;
 import org.hisp.dhis.system.database.DatabaseInfo;
@@ -94,8 +93,7 @@ public class JdbcCompletenessTargetTableManager extends AbstractJdbcTableManager
       PartitionManager partitionManager,
       DatabaseInfo databaseInfo,
       JdbcTemplate jdbcTemplate,
-      AnalyticsExportSettings analyticsExportSettings,
-      PeriodDataProvider periodDataProvider) {
+      AnalyticsExportSettings analyticsExportSettings) {
     super(
         idObjectManager,
         organisationUnitService,
@@ -108,8 +106,7 @@ public class JdbcCompletenessTargetTableManager extends AbstractJdbcTableManager
         partitionManager,
         databaseInfo,
         jdbcTemplate,
-        analyticsExportSettings,
-        periodDataProvider);
+        analyticsExportSettings);
   }
 
   @Override
