+++ dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/table/JdbcValidationResultTableManager.java
@@ -61,7 +61,6 @@ import org.hisp.dhis.jdbc.StatementBuilder;
 import org.hisp.dhis.organisationunit.OrganisationUnitGroupSet;
 import org.hisp.dhis.organisationunit.OrganisationUnitLevel;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.period.PeriodDataProvider;
 import org.hisp.dhis.period.PeriodType;
 import org.hisp.dhis.resourcetable.ResourceTableService;
 import org.hisp.dhis.setting.SystemSettingManager;
@@ -94,8 +93,7 @@ public class JdbcValidationResultTableManager extends AbstractJdbcTableManager {
       PartitionManager partitionManager,
       DatabaseInfo databaseInfo,
       JdbcTemplate jdbcTemplate,
-      AnalyticsExportSettings analyticsExportSettings,
-      PeriodDataProvider periodDataProvider) {
+      AnalyticsExportSettings analyticsExportSettings) {
     super(
         idObjectManager,
         organisationUnitService,
@@ -108,8 +106,7 @@ public class JdbcValidationResultTableManager extends AbstractJdbcTableManager {
         partitionManager,
         databaseInfo,
         jdbcTemplate,
-        analyticsExportSettings,
-        periodDataProvider);
+        analyticsExportSettings);
   }
 
   @Override
