+++ dhis-2/dhis-services/dhis-service-analytics/src/test/java/org/hisp/dhis/analytics/table/JdbcAnalyticsTableManagerTest.java
@@ -51,7 +51,6 @@ import org.hisp.dhis.common.IdentifiableObjectManager;
 import org.hisp.dhis.dataapproval.DataApprovalLevelService;
 import org.hisp.dhis.jdbc.StatementBuilder;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.period.PeriodDataProvider;
 import org.hisp.dhis.resourcetable.ResourceTableService;
 import org.hisp.dhis.setting.SettingKey;
 import org.hisp.dhis.setting.SystemSettingManager;
@@ -81,8 +80,6 @@ class JdbcAnalyticsTableManagerTest {
 
   @Mock private AnalyticsExportSettings analyticsExportSettings;
 
-  @Mock private PeriodDataProvider periodDataProvider;
-
   private AnalyticsTableManager subject;
 
   @BeforeEach
@@ -100,8 +97,7 @@ class JdbcAnalyticsTableManagerTest {
             mock(PartitionManager.class),
             mock(DatabaseInfo.class),
             jdbcTemplate,
-            analyticsExportSettings,
-            periodDataProvider);
+            analyticsExportSettings);
   }
 
   @Test
