+++ dhis-2/dhis-services/dhis-service-analytics/src/test/java/org/hisp/dhis/analytics/table/JdbcEnrollmentAnalyticsTableManagerTest.java
@@ -48,7 +48,6 @@ import org.hisp.dhis.common.ValueType;
 import org.hisp.dhis.dataapproval.DataApprovalLevelService;
 import org.hisp.dhis.jdbc.statementbuilder.PostgreSQLStatementBuilder;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.period.PeriodDataProvider;
 import org.hisp.dhis.program.Program;
 import org.hisp.dhis.program.ProgramTrackedEntityAttribute;
 import org.hisp.dhis.resourcetable.ResourceTableService;
@@ -78,8 +77,6 @@ class JdbcEnrollmentAnalyticsTableManagerTest {
 
   @Mock private AnalyticsExportSettings analyticsExportSettings;
 
-  @Mock private PeriodDataProvider periodDataProvider;
-
   private JdbcEnrollmentAnalyticsTableManager subject;
 
   private static final Date START_TIME = new DateTime(2019, 8, 1, 0, 0).toDate();
@@ -99,8 +96,7 @@ class JdbcEnrollmentAnalyticsTableManagerTest {
             mock(PartitionManager.class),
             databaseInfo,
             jdbcTemplate,
-            analyticsExportSettings,
-            periodDataProvider);
+            analyticsExportSettings);
   }
 
   @Test
