+++ dhis-2/dhis-services/dhis-service-analytics/src/test/java/org/hisp/dhis/analytics/table/JdbcEventAnalyticsTableManagerTest.java
@@ -27,7 +27,6 @@
  */
 package org.hisp.dhis.analytics.table;
 
-import static java.time.LocalDate.now;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
@@ -50,6 +49,7 @@ import static org.hisp.dhis.analytics.ColumnDataType.GEOMETRY_POINT;
 import static org.hisp.dhis.analytics.ColumnDataType.INTEGER;
 import static org.hisp.dhis.analytics.ColumnDataType.TEXT;
 import static org.hisp.dhis.analytics.ColumnDataType.TIMESTAMP;
+import static org.hisp.dhis.resourcetable.ResourceTable.LATEST_YEAR_SUPPORTED;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -92,7 +92,6 @@ import org.hisp.dhis.jdbc.statementbuilder.PostgreSQLStatementBuilder;
 import org.hisp.dhis.organisationunit.OrganisationUnitGroupSet;
 import org.hisp.dhis.organisationunit.OrganisationUnitLevel;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.period.PeriodDataProvider;
 import org.hisp.dhis.period.PeriodType;
 import org.hisp.dhis.program.Program;
 import org.hisp.dhis.program.ProgramStage;
@@ -135,8 +134,6 @@ class JdbcEventAnalyticsTableManagerTest {
 
   @Mock private AnalyticsExportSettings analyticsExportSettings;
 
-  @Mock private PeriodDataProvider periodDataProvider;
-
   private JdbcEventAnalyticsTableManager subject;
 
   private Date today;
@@ -178,8 +175,7 @@ class JdbcEventAnalyticsTableManagerTest {
             mock(PartitionManager.class),
             databaseInfo,
             jdbcTemplate,
-            analyticsExportSettings,
-            periodDataProvider);
+            analyticsExportSettings);
   }
 
   @Test
@@ -220,7 +216,6 @@ class JdbcEventAnalyticsTableManagerTest {
         .thenReturn(lastLatestPartitionUpdate);
     when(jdbcTemplate.queryForList(Mockito.anyString())).thenReturn(queryResp);
     when(idObjectManager.getAllNoAcl(Program.class)).thenReturn(programs);
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
 
     List<AnalyticsTable> tables = subject.getAnalyticsTables(params);
     assertThat(tables, hasSize(2));
@@ -258,12 +253,8 @@ class JdbcEventAnalyticsTableManagerTest {
     addCategoryCombo(program, categoryCombo);
 
     when(idObjectManager.getAllNoAcl(Program.class)).thenReturn(Lists.newArrayList(program));
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
 
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
-
-    when(jdbcTemplate.queryForList(
-            getYearQueryForCurrentYear(program, true, availableDataYears), Integer.class))
+    when(jdbcTemplate.queryForList(getYearQueryForCurrentYear(program, true), Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
     AnalyticsTableUpdateParams params =
@@ -280,7 +271,7 @@ class JdbcEventAnalyticsTableManagerTest {
     new AnalyticsTableAsserter.Builder(tables.get(0))
         .withTableType(AnalyticsTableType.EVENT)
         .withTableName(TABLE_PREFIX + program.getUid().toLowerCase())
-        .withColumnSize(56)
+        .withColumnSize(55)
         .withDefaultColumns(subject.getFixedColumns())
         .addColumns(periodColumns)
         .addColumn(categoryA.getUid(), CHARACTER_11, "acs.", categoryA.getCreated())
@@ -307,7 +298,6 @@ class JdbcEventAnalyticsTableManagerTest {
     program.setProgramStages(Sets.newHashSet(ps1));
 
     when(idObjectManager.getAllNoAcl(Program.class)).thenReturn(Lists.newArrayList(program));
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
 
     String aliasD1 = "(select eventdatavalues #>> '{%s, value}' " + FROM_CLAUSE + " ) as \"%s\"";
     String aliasD2 =
@@ -369,10 +359,7 @@ class JdbcEventAnalyticsTableManagerTest {
             .withToday(today)
             .build();
 
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
-
-    when(jdbcTemplate.queryForList(
-            getYearQueryForCurrentYear(program, true, availableDataYears), Integer.class))
+    when(jdbcTemplate.queryForList(getYearQueryForCurrentYear(program, true), Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
     List<AnalyticsTable> tables = subject.getAnalyticsTables(params);
@@ -382,7 +369,7 @@ class JdbcEventAnalyticsTableManagerTest {
     new AnalyticsTableAsserter.Builder(tables.get(0))
         .withTableName(TABLE_PREFIX + program.getUid().toLowerCase())
         .withTableType(AnalyticsTableType.EVENT)
-        .withColumnSize(63)
+        .withColumnSize(62)
         .addColumns(periodColumns)
         .addColumn(d1.getUid(), TEXT, toAlias(aliasD1, d1.getUid())) // ValueType.TEXT
         .addColumn(d2.getUid(), DOUBLE, toAlias(aliasD2, d2.getUid())) // ValueType.PERCENTAGE
@@ -421,9 +408,6 @@ class JdbcEventAnalyticsTableManagerTest {
     program.setProgramStages(Sets.newHashSet(ps1));
 
     when(idObjectManager.getAllNoAcl(Program.class)).thenReturn(Lists.newArrayList(program));
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
-
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
 
     String aliasD1 = "(select eventdatavalues #>> '{%s, value}' " + FROM_CLAUSE + " ) as \"%s\"";
     String aliasTea1 =
@@ -438,8 +422,7 @@ class JdbcEventAnalyticsTableManagerTest {
             .withToday(today)
             .build();
 
-    when(jdbcTemplate.queryForList(
-            getYearQueryForCurrentYear(program, true, availableDataYears), Integer.class))
+    when(jdbcTemplate.queryForList(getYearQueryForCurrentYear(program, true), Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
     List<AnalyticsTable> tables = subject.getAnalyticsTables(params);
@@ -449,7 +432,7 @@ class JdbcEventAnalyticsTableManagerTest {
     new AnalyticsTableAsserter.Builder(tables.get(0))
         .withTableName(TABLE_PREFIX + program.getUid().toLowerCase())
         .withTableType(AnalyticsTableType.EVENT)
-        .withColumnSize(58)
+        .withColumnSize(57)
         .addColumns(periodColumns)
         .addColumn(d1.getUid(), TEXT, toAlias(aliasD1, d1.getUid())) // ValueType.TEXT
         .addColumn(
@@ -482,9 +465,6 @@ class JdbcEventAnalyticsTableManagerTest {
     programA.setProgramStages(Sets.newHashSet(ps1));
 
     when(idObjectManager.getAllNoAcl(Program.class)).thenReturn(Lists.newArrayList(programA));
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
-
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
 
     AnalyticsTableUpdateParams params =
         AnalyticsTableUpdateParams.newBuilder()
@@ -493,8 +473,7 @@ class JdbcEventAnalyticsTableManagerTest {
             .withToday(today)
             .build();
 
-    when(jdbcTemplate.queryForList(
-            getYearQueryForCurrentYear(programA, true, availableDataYears), Integer.class))
+    when(jdbcTemplate.queryForList(getYearQueryForCurrentYear(programA, true), Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
     subject.populateTable(
@@ -530,9 +509,6 @@ class JdbcEventAnalyticsTableManagerTest {
     programA.setProgramAttributes(Lists.newArrayList(programTrackedEntityAttribute));
 
     when(idObjectManager.getAllNoAcl(Program.class)).thenReturn(Lists.newArrayList(programA));
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
-
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
 
     AnalyticsTableUpdateParams params =
         AnalyticsTableUpdateParams.newBuilder()
@@ -541,8 +517,7 @@ class JdbcEventAnalyticsTableManagerTest {
             .withToday(today)
             .build();
 
-    when(jdbcTemplate.queryForList(
-            getYearQueryForCurrentYear(programA, true, availableDataYears), Integer.class))
+    when(jdbcTemplate.queryForList(getYearQueryForCurrentYear(programA, true), Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
     subject.populateTable(
@@ -577,9 +552,6 @@ class JdbcEventAnalyticsTableManagerTest {
     programA.setProgramAttributes(Lists.newArrayList(programTrackedEntityAttribute));
 
     when(idObjectManager.getAllNoAcl(Program.class)).thenReturn(Lists.newArrayList(programA));
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
-
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
 
     AnalyticsTableUpdateParams params =
         AnalyticsTableUpdateParams.newBuilder()
@@ -588,8 +560,7 @@ class JdbcEventAnalyticsTableManagerTest {
             .withToday(today)
             .build();
 
-    when(jdbcTemplate.queryForList(
-            getYearQueryForCurrentYear(programA, true, availableDataYears), Integer.class))
+    when(jdbcTemplate.queryForList(getYearQueryForCurrentYear(programA, true), Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
     // When
@@ -614,11 +585,6 @@ class JdbcEventAnalyticsTableManagerTest {
         rnd.objects(OrganisationUnitLevel.class, 2).collect(Collectors.toList());
     Program programA = rnd.nextObject(Program.class);
     programA.setId(0);
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
-
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
-    int startYear = availableDataYears.get(0);
-    int latestYear = availableDataYears.get(availableDataYears.size() - 1);
 
     when(idObjectManager.getAllNoAcl(Program.class))
         .thenReturn(Collections.singletonList(programA));
@@ -634,10 +600,8 @@ class JdbcEventAnalyticsTableManagerTest {
                 + "and ("
                 + getDateLinkedToStatus()
                 + ") > '1000-01-01' and psi.deleted is false ) "
-                + "as temp where temp.supportedyear >= "
-                + startYear
-                + " and temp.supportedyear <= "
-                + latestYear,
+                + "as temp where temp.supportedyear >= 1975 and temp.supportedyear <= "
+                + LATEST_YEAR_SUPPORTED,
             Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
@@ -677,14 +641,10 @@ class JdbcEventAnalyticsTableManagerTest {
         .thenReturn(Collections.singletonList(programA));
     when(idObjectManager.getDataDimensionsNoAcl(OrganisationUnitGroupSet.class))
         .thenReturn(ouGroupSet);
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
-
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
 
     AnalyticsTableUpdateParams params =
         AnalyticsTableUpdateParams.newBuilder().withStartTime(START_TIME).build();
-    when(jdbcTemplate.queryForList(
-            getYearQueryForCurrentYear(programA, false, availableDataYears), Integer.class))
+    when(jdbcTemplate.queryForList(getYearQueryForCurrentYear(programA, false), Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
     List<AnalyticsTable> tables = subject.getAnalyticsTables(params);
@@ -717,12 +677,7 @@ class JdbcEventAnalyticsTableManagerTest {
     when(idObjectManager.getAllNoAcl(Program.class))
         .thenReturn(Collections.singletonList(programA));
     when(categoryService.getAttributeCategoryOptionGroupSetsNoAcl()).thenReturn(cogs);
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
-
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
-
-    when(jdbcTemplate.queryForList(
-            getYearQueryForCurrentYear(programA, false, availableDataYears), Integer.class))
+    when(jdbcTemplate.queryForList(getYearQueryForCurrentYear(programA, false), Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
     AnalyticsTableUpdateParams params =
@@ -796,11 +751,6 @@ class JdbcEventAnalyticsTableManagerTest {
     programA.setProgramAttributes(Lists.newArrayList(programTrackedEntityAttribute));
 
     when(idObjectManager.getAllNoAcl(Program.class)).thenReturn(Lists.newArrayList(programA));
-    when(periodDataProvider.getAvailableYears()).thenReturn(List.of(2018, 2019, now().getYear()));
-
-    List<Integer> availableDataYears = periodDataProvider.getAvailableYears();
-    int startYear = availableDataYears.get(0);
-    int latestYear = availableDataYears.get(availableDataYears.size() - 1);
 
     when(jdbcTemplate.queryForList(
             "select temp.supportedyear from (select distinct extract(year from "
@@ -815,10 +765,8 @@ class JdbcEventAnalyticsTableManagerTest {
                 + ") > '1000-01-01' and psi.deleted is false and ("
                 + getDateLinkedToStatus()
                 + ") >= '2018-01-01') "
-                + "as temp where temp.supportedyear >= "
-                + startYear
-                + " and temp.supportedyear <= "
-                + latestYear,
+                + "as temp where temp.supportedyear >= 1975 and temp.supportedyear <= "
+                + LATEST_YEAR_SUPPORTED,
             Integer.class))
         .thenReturn(Lists.newArrayList(2018, 2019));
 
@@ -853,10 +801,7 @@ class JdbcEventAnalyticsTableManagerTest {
     program.setCategoryCombo(categoryCombo);
   }
 
-  private String getYearQueryForCurrentYear(
-      Program program, boolean withExecutionDate, List<Integer> availableDataYears) {
-    int startYear = availableDataYears.get(0);
-    int latestYear = availableDataYears.get(availableDataYears.size() - 1);
+  private String getYearQueryForCurrentYear(Program program, boolean withExecutionDate) {
     String sql =
         "select temp.supportedyear from (select distinct "
             + "extract(year from "
@@ -877,10 +822,8 @@ class JdbcEventAnalyticsTableManagerTest {
     }
 
     sql +=
-        ") as temp where temp.supportedyear >= "
-            + startYear
-            + " and temp.supportedyear <= "
-            + latestYear;
+        ") as temp where temp.supportedyear >= 1975 and temp.supportedyear <= "
+            + LATEST_YEAR_SUPPORTED;
 
     return sql;
   }
