+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/i18n_global.properties
@@ -603,7 +603,6 @@ BiWeekly=Bi-weekly
 Monthly=Monthly
 BiMonthly=Bi-monthly
 Quarterly=Quarterly
-QuarterlyNov=Quarterly November
 SixMonthly=Six-monthly
 SixMonthlyApril=Six-monthly April
 SixMonthlyNov=Six-monthly November
@@ -640,8 +639,6 @@ format.BiMonthly.startDate=MMM 'to '
 format.BiMonthly.endDate=MMM yyyy
 format.Quarterly.startDate=MMM 'to '
 format.Quarterly.endDate=MMM yyyy
-format.QuarterlyNov.startDate=MMM 'to '
-format.QuarterlyNov.endDate=MMM yyyy
 format.SixMonthly.startDate=MMM 'to '
 format.SixMonthly.endDate=MMM yyyy
 format.SixMonthlyApril.startDate=MMM 'to '
@@ -732,53 +729,9 @@ weekday.short.friday=Fri
 weekday.short.saturday=Sat
 weekday.short.sunday=Sun
 
-# Various Calendar months.
-# The keys for months are still based on ISO 8601, the ordering is correct, but the key names are not. Jan-Dec -> 1-12. #
-
-#-- Ethiopian Calendar Months and days -----------------------------------------------------------#
-ethiopian.month.january=Meskerem
-ethiopian.month.february=Tikemet
-ethiopian.month.march=Hidar
-ethiopian.month.april=Tahesas
-ethiopian.month.may=Tir
-ethiopian.month.june=Yekatit
-ethiopian.month.july=Megabit
-ethiopian.month.august=Miazia
-ethiopian.month.september=Ginbot
-ethiopian.month.october=Sene
-ethiopian.month.november=Hamle
-ethiopian.month.december=Nehase.
-
-ethiopian.month.short.january=Mes
-ethiopian.month.short.february=Tik
-ethiopian.month.short.march=Hid
-ethiopian.month.short.april=Tah
-ethiopian.month.short.may=Tir
-ethiopian.month.short.june=Yek
-ethiopian.month.short.july=Meg
-ethiopian.month.short.august=Mia
-ethiopian.month.short.september=Gin
-ethiopian.month.short.october=Sen
-ethiopian.month.short.november=Ham
-ethiopian.month.short.december=Neh
-
-ethiopian.weekday.monday=Segno
-ethiopian.weekday.tuesday=Maksegno
-ethiopian.weekday.wednesday=Irob
-ethiopian.weekday.thursday=Hamus
-ethiopian.weekday.friday=Arb
-ethiopian.weekday.saturday=Kidame
-ethiopian.weekday.sunday=Ehud
-
-ethiopian.weekday.short.monday=Seg
-ethiopian.weekday.short.tuesday=Mak
-ethiopian.weekday.short.wednesday=Iro
-ethiopian.weekday.short.thursday=Ham
-ethiopian.weekday.short.friday=Arb
-ethiopian.weekday.short.saturday=Kid
-ethiopian.weekday.short.sunday=Ehu
-
 #-- Nepali Calendar Month/Day names -------------------------------------------#
+
+# The keys for months are still based on ISO 8601, the ordering is correct, but the key names are not. Jan-Dec -> 1-12. #
 nepali.month.january=Baishak
 nepali.month.february=Jestha
 nepali.month.march=Asar
