+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/org/hisp/dhis/period/hibernate/PeriodType.hbm.xml
@@ -24,7 +24,6 @@
     <subclass name="org.hisp.dhis.period.MonthlyPeriodType" discriminator-value="Monthly" />
     <subclass name="org.hisp.dhis.period.BiMonthlyPeriodType" discriminator-value="BiMonthly" />
     <subclass name="org.hisp.dhis.period.QuarterlyPeriodType" discriminator-value="Quarterly" />
-    <subclass name="org.hisp.dhis.period.QuarterlyNovemberPeriodType" discriminator-value="QuarterlyNov" />
     <subclass name="org.hisp.dhis.period.SixMonthlyPeriodType" discriminator-value="SixMonthly" />
     <subclass name="org.hisp.dhis.period.SixMonthlyAprilPeriodType" discriminator-value="SixMonthlyApril" />
     <subclass name="org.hisp.dhis.period.SixMonthlyNovemberPeriodType" discriminator-value="SixMonthlyNov" />
@@ -37,4 +36,3 @@
 
   </class>
 </hibernate-mapping>
-8
